Use CLI flag instead of config object
