+++ lib/spinner.js
@@ -10,16 +10,6 @@ exports.create = message => {
 	global.spinner = ora(message).start();
 };
 
-exports.succeed = () => {
-	if (global.spinner) {
-		global.spinner.succeed();
-
-		// Prevents the spinner from getting succeeded
-		// again once new spinner gets created
-		global.spinner = null;
-	}
-};
-
 exports.fail = message => {
 	if (global.spinner) {
 		global.spinner.fail();
